[change] 別ファイルに切り出しました。
